# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Momentum Terminal API",
    "description": "US equities 12-1 momentum research and INTERNAL paper simulation. No brokerage connectivity.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/universe": {
      "get": {
        "tags": [
          "universe"
        ],
        "summary": "Universe",
        "operationId": "universe_api_universe_get",
        "parameters": [
          {
            "name": "session",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Signal session (default: latest stored session)",
              "title": "Session"
            },
            "description": "Signal session (default: latest stored session)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UniverseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/universe/{symbol}": {
      "get": {
        "tags": [
          "universe"
        ],
        "summary": "Stock Detail",
        "operationId": "stock_detail_api_universe__symbol__get",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "session",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CorporateActionModel": {
        "properties": {
          "ex_date": {
            "type": "string",
            "title": "Ex Date"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ratio"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "ex_date",
          "action_type",
          "ratio",
          "amount"
        ],
        "title": "CorporateActionModel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PricePoint": {
        "properties": {
          "session": {
            "type": "string",
            "title": "Session"
          },
          "open": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open"
          },
          "high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "High"
          },
          "low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low"
          },
          "close": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "tr_index": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tr Index"
          }
        },
        "type": "object",
        "required": [
          "session",
          "open",
          "high",
          "low",
          "close",
          "volume",
          "tr_index"
        ],
        "title": "PricePoint"
      },
...
      "SignalInputs": {
        "properties": {
          "signal_session": {
            "type": "string",
            "title": "Signal Session"
          },
          "session_t21": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session T21"
          },
          "session_t252": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session T252"
          },
          "close_t": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close T"
          },
          "close_t21_raw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close T21 Raw"
          },
          "close_t252_raw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close T252 Raw"
          },
          "tr_t21": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tr T21"
          },
          "tr_t252": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tr T252"
          },
          "momentum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Momentum"
          },
          "adv20": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adv20"
          },
          "valid_history": {
            "type": "integer",
            "title": "Valid History"
          },
          "eligible": {
            "type": "boolean",
            "title": "Eligible"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "reason_text": {
            "type": "string",
            "title": "Reason Text"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "formula": {
            "type": "string",
            "title": "Formula"
          }
        },
        "type": "object",
        "required": [
          "signal_session",
          "session_t21",
          "session_t252",
          "close_t",
          "close_t21_raw",
          "close_t252_raw",
          "tr_t21",
          "tr_t252",
          "momentum",
          "adv20",
          "valid_history",
          "eligible",
          "reason",
          "reason_text",
          "rank",
          "formula"
        ],
        "title": "SignalInputs"
      },
...
      "StockDetail": {
        "properties": {
          "data_label": {
            "type": "string",
            "title": "Data Label"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "sector_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector Source"
          },
          "asset_type": {
            "type": "string",
            "title": "Asset Type"
          },
          "asset_type_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Type Source"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "list_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "List Date"
          },
          "delist_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delist Date"
          },
          "signal": {
            "$ref": "#/components/schemas/SignalInputs"
          },
          "prices": {
            "items": {
              "$ref": "#/components/schemas/PricePoint"
            },
            "type": "array",
            "title": "Prices"
          },
          "corporate_actions": {
            "items": {
              "$ref": "#/components/schemas/CorporateActionModel"
            },
            "type": "array",
            "title": "Corporate Actions"
          },
          "position_shares": {
            "type": "integer",
            "title": "Position Shares"
          }
        },
        "type": "object",
        "required": [
          "data_label",
          "symbol",
          "name",
          "sector",
          "sector_source",
          "asset_type",
          "asset_type_source",
          "exchange",
          "list_date",
          "delist_date",
          "signal",
          "prices",
          "corporate_actions",
          "position_shares"
        ],
        "title": "StockDetail"
      },
      "StrategyConfig": {
        "properties": {
          "initial_capital": {
            "type": "number",
            "maximum": 10000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Initial Capital",
            "description": "Starting virtual cash, USD",
            "default": 100000.0
          },
          "top_n": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Top N",
            "description": "Number of top-ranked eligible stocks to hold",
            "default": 50
          },
          "lookback_sessions": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 2.0,
            "title": "Lookback Sessions",
            "description": "Older momentum anchor: sessions before signal",
            "default": 252
          },
          "skip_sessions": {
            "type": "integer",
            "maximum": 252.0,
            "minimum": 0.0,
            "title": "Skip Sessions",
            "description": "Recent anchor: sessions before signal (skip month)",
            "default": 21
          },
          "min_history_sessions": {
            "type": "integer",
            "maximum": 2000.0,
            "minimum": 0.0,
            "title": "Min History Sessions",
            "description": "Valid bars required before the signal session",
            "default": 252
          },
          "min_price": {
            "type": "number",
            "minimum": 0.0,
            "title": "Min Price",
            "description": "Raw (unadjusted) close at the signal session must exceed this, USD",
            "default": 5.0
          },
          "adv_window": {
            "type": "integer",
            "maximum": 252.0,
            "minimum": 1.0,
            "title": "Adv Window",
            "description": "Sessions in the average dollar-volume window",
            "default": 20
          },
          "min_adv_usd": {
            "type": "number",
            "minimum": 0.0,
            "title": "Min Adv Usd",
            "description": "Minimum trailing average daily dollar volume, USD",
            "default": 5000000.0
          },
          "slippage_bps": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Slippage Bps",
            "description": "Assumed adverse slippage vs. the open, basis points",
            "default": 10.0
          },
          "commission_per_order": {
            "type": "number",
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Commission Per Order",
            "description": "Assumed flat commission per fill, USD",
            "default": 0.0
          },
          "commission_bps": {
            "type": "number",
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Commission Bps",
            "description": "Assumed commission as bps of traded value",
            "default": 0.0
          },
          "min_session_coverage": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Session Coverage",
            "description": "Share of the point-in-time universe that must have a bar at the signal session; otherwise the rebalance is blocked",
            "default": 0.9
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date",
            "description": "Backtest start (ISO date); ignored by the paper ledger"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date",
            "description": "Backtest end (ISO date); ignored by the paper ledger"
          },
          "benchmark_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Benchmark Symbol",
            "description": "Benchmark; default = provider's SPY (or demo stand-in)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "StrategyConfig",
        "description": "All tunable inputs of the 12-1 momentum strategy.\n\nDefaults are the documented v1 specification. Every backtest run and every\npaper-portfolio configuration version stores the full JSON of this model."
      },
...
      "UniverseResponse": {
        "properties": {
          "data_label": {
            "type": "string",
            "title": "Data Label"
          },
          "signal_session": {
            "type": "string",
            "title": "Signal Session"
          },
          "is_month_end": {
            "type": "boolean",
            "title": "Is Month End"
          },
          "basis_note": {
            "type": "string",
            "title": "Basis Note"
          },
          "plan_signal_session": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Signal Session"
          },
          "portfolio_as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio As Of"
          },
          "config": {
            "$ref": "#/components/schemas/StrategyConfig"
          },
          "universe_count": {
            "type": "integer",
            "title": "Universe Count"
          },
          "eligible_count": {
            "type": "integer",
            "title": "Eligible Count"
          },
          "selected_count": {
            "type": "integer",
            "title": "Selected Count"
          },
          "coverage": {
            "type": "number",
            "title": "Coverage"
          },
          "blocked_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blocked Reason"
          },
          "reason_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Reason Counts"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/UniverseRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "data_label",
          "signal_session",
          "is_month_end",
          "basis_note",
          "plan_signal_session",
          "portfolio_as_of",
          "config",
          "universe_count",
          "eligible_count",
          "selected_count",
          "coverage",
          "blocked_reason",
          "reason_counts",
          "rows"
        ],
        "title": "UniverseResponse"
      },
      "UniverseRow": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "asset_type": {
            "type": "string",
            "title": "Asset Type"
          },
          "close_raw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Raw"
          },
          "adv20": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adv20"
          },
          "momentum": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Momentum"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "eligible": {
            "type": "boolean",
            "title": "Eligible"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "reason_text": {
            "type": "string",
            "title": "Reason Text"
          },
          "selected": {
            "type": "boolean",
            "title": "Selected"
          },
          "model_weight": {
            "type": "number",
            "title": "Model Weight",
            "description": "Equal weight if selected by this live ranking"
          },
          "plan_target_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Target Weight",
            "description": "Target weight in the latest frozen paper plan"
          },
          "position_shares": {
            "type": "integer",
            "title": "Position Shares"
          },
          "current_weight": {
            "type": "number",
            "title": "Current Weight"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name",
          "sector",
          "asset_type",
          "close_raw",
          "adv20",
          "momentum",
          "rank",
          "eligible",
          "reason",
          "reason_text",
          "selected",
          "model_weight",
          "plan_target_weight",
          "position_shares",
          "current_weight"
        ],
        "title": "UniverseRow"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}